Add a timer function to BrainstormViewModel.
Motivation: The user need a feedback for the time amount left in his brainstorming session.

Modifications: In the ViewModel file BrainstormingViewModel.swif, a function called startBrainstormTimer is added. This function is called when you want to start a timer with the amount of time that is passed. The timer also updates the timer String that will be used as the data for the label in the UI.
Also it was added a variable that instatiates the timer in the ViewModel

Results: A timer is now created so the it can be used to inform the user the amount of time left.