Format the timer in minutes and seconds
Motivation: Provide the user a way to keep track of the time closer to his reality

Modifications: In the previous timer that was created, it was added a few conditionals checks to format the timer in minutes and seconds. The conditionals are to keep a standard of two digits numbers in the timer display.

Result: The timer now works with minutes as well and format their string so it can be displayed in the label in a correct way.