Add the ViewModels to the iOS side of app Motivation: Integrate the data so it can be easier to manipulate with the view.
Modifications: The existing ViewModels in the iOS side of the app were modified so they could work propperly to provide the correct data to the View. Also it was created OSLog class to keep track of the errors and handle them. An Idea Struct was created to make data manipulation easier.

Results: Data is now easier to manipulate