Make it possible to reset Brainstorm cycle through RankingView
Motivation: To give the possibility to our users to restart their Brainstorm cycles as many times as possible without restarting the app.

Modifications: Changed the the restart Button in the RankingView to a NavigationLink button that instatiate the MenuView again, making it so our app has a full cycle.

Result: The Brainstorming cycle can be restarted now without closing and openning the app.